requirements.txt
# PDF Merger Pro - Зависимости

# Основные зависимости
PyQt6>=6.5.0
PyPDF2>=3.0.0
qtawesome>=1.2.0

# Опциональные зависимости для предварительного просмотра
PyMuPDF>=1.23.0

# Зависимости для конвертации файлов
Pillow>=10.0.0          # Для конвертации изображений
reportlab>=4.0.0        # Для создания PDF из текста и изображений
docx2pdf>=0.1.8         # Для конвертации Word документов

# Для разработки (опционально)
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0
